pyproject.toml
[tool.poetry]
name = "madr"
version = "0.1.0"
description = ""
authors = ["Rafael Santos <rafaelsantosdearaujo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
psycopg2 = "^2.9.10"
psycopg = "^3.2.3"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.2"
pytest = "^8.3.3"
taskipy = "^1.14.0"
pytest-cov = "^6.0.0"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
pre_test = 'task lint'
test = 'pytest -s -x --cov=madr -vv'
post_test = 'coverage html'
run = 'fastapi dev madr/app.py'
auto_migration = 'alembic revision --autogenerate -m'
migration_up = 'alembic upgrade head'
migration_down = 'alembic downgrade -1'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
